+++ renderer/src/modules/webpackmodules.js
@@ -142,8 +142,8 @@ export default class WebpackModules {
 
     /**
      * Finds a module using a filter function.
-     * @param {Function} filter A function to use to filter modules
-     * @param {object} [options] Whether to return only the first matching module
+     * @param {function} filter A function to use to filter modules
+     * @param {object} [options] Set of options to customize the search
      * @param {Boolean} [options.first=true] Whether to return only the first matching module
      * @param {Boolean} [options.defaultExport=true] Whether to return default export when matching the default export
      * @return {Any}
@@ -312,7 +312,7 @@ export default class WebpackModules {
     /**
      * Finds a module that lazily loaded.
      * @param {(m) => boolean} filter A function to use to filter modules.
-     * @param {object} [options] Whether to return only the first matching module
+     * @param {object} [options] Set of options to customize the search
      * @param {AbortSignal} [options.signal] AbortSignal of an AbortController to cancel the promise
      * @param {Boolean} [options.defaultExport=true] Whether to return default export when matching the default export
      * @returns {Promise<any>}
@@ -320,25 +320,25 @@ export default class WebpackModules {
     static getLazy(filter, options = {}) {
         /** @type {AbortSignal} */
         const abortSignal = options.signal;
-        const defaultExport = options.defaultExport;
+        const defaultExport = options.defaultExport ?? true;
         const fromCache = this.getModule(filter);
         if (fromCache) return Promise.resolve(fromCache);
 
         return new Promise((resolve) => {
             const cancel = () => {this.removeListener(listener);};
-            const listener = function (m) {
-                const directMatch = filter(m);
+            const listener = function (mod) {
+                const directMatch = filter(mod);
                 
                 if (directMatch) {
                     cancel();
                     return resolve(directMatch);
                 }
 
-                const defaultMatch = filter(m.default);
+                const defaultMatch = filter(mod.default);
                 if (!defaultMatch) return; 
 
                 cancel();
-                resolve(defaultExport ? m.default : defaultExport);
+                resolve(defaultExport ? mod.default : mod);
             };
 
             this.addListener(listener);