+++ renderer/src/structs/metaerror.js
@@ -1,6 +1,0 @@
-export default class MetaError extends Error {
-    constructor(message) {
-        super(message);
-        this.name = "MetaError";
-    }
-}