+++ CHANGELOG.md
@@ -2,6 +2,19 @@
 
 This changelog starts with the restructured 1.0.0 release that happened after context isolation changes. The changelogs here should more-or-less mirror the ones that get shown in the client but probably with less formatting and pizzazz.
 
+## 1.6.1
+
+### Added
+
+### Removed
+
+### Changed
+
+### Fixed
+- Fixed an issue where `waitForModule` would not return the found module.
+- Fixed an issue where broken addon METAs could prevent BD from fully loading.
+- Fixed an issue where developer badges stopped rendering.
+
 ## 1.6.0
 
 ### Added